Move termination invariant checks into separate module
