+++ rust/utils/run-locally/src/invariants/common.rs
@@ -1,0 +1,3 @@
+// This number should be even, so the messages can be split into two equal halves
+// sent before and after the relayer spins up, to avoid rounding errors.
+pub const SOL_MESSAGES_EXPECTED: u32 = 20;
